+++ src/Phan/AST/Visitor/KindVisitor.php
@@ -317,6 +317,11 @@ interface KindVisitor
     public function visitType(Node $node);
 
     /**
+     * Visit a node with kind `\ast\AST_TYPE_UNION`
+     */
+    public function visitTypeUnion(Node $node);
+
+    /**
      * Visit a node with kind `\ast\AST_NULLABLE_TYPE`
      */
     public function visitNullableType(Node $node);